+++ test/EFCore.Cosmos.FunctionalTests/Query/SimpleQueryCosmosTest.cs
@@ -1799,7 +1799,7 @@ WHERE (c[""Discriminator""] = ""Product"")");
                       .Where(
                         p => od
                             .Where(o => o.ProductID == p.ProductID)
-                            .Select(odd => odd.Quantity).Contains<short>(5)),
+                            .Select(odd => odd.Quantity).Contains<int>(5)),
                 entryCount: 1);
 
             AssertSql(
